pyproject.toml
[tool.poetry]
name = "bigdata-hw1"
version = "0.1.0"
description = ""
authors = ["Aviad Hadad <aviad@sthuck.info>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
cassandra-driver = "^3.29.0"
pandas = "^2.2.0"
duckdb = "^0.9.2"
plotly = "^5.18.0"
geopandas = "^0.14.3"
plotly-geo = "^1.0.0"
pyshp = "^2.3.1"
shapely = "^2.0.2"
matplotlib = "^3.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
